grpc 1.60 patch for wolfSSL support
This patch implements wolfSSL support in grpc with CMake.

Compile wolfSSL with:
  ./configure --enable-all 'CPPFLAGS=-DWOLFSSL_RSA_KEY_CHECK -DHAVE_EX_DATA_CLEANUP_HOOKS'
  make
  make install

Compile grpc with:
  mkdir cmake/build
  cd cmake/build
  cmake -DgRPC_SSL_PROVIDER=wolfssl ../..
  make
  make install

For a custom wolfSSL install directory use:
  cmake -DgRPC_SSL_PROVIDER=wolfssl -DWOLFSSL_INSTALL_DIR=<path/to/install/dir> ../..

The following tests pass with wolfSSL:
  bad_ssl_alpn_test bad_ssl_cert_test client_ssl_test crl_ssl_transport_security_test server_ssl_test ssl_transport_security_test ssl_transport_security_utils_test test_core_security_ssl_credentials_test test_cpp_end2end_ssl_credentials_test h2_ssl_cert_test h2_ssl_session_reuse_test

The tests require that the loopback interface has both IPv4 and IPv6 enabled.

The tests need to be run from the root directory
  ./cmake/build/ssl_transport_security_test

To build grpc for debugging use:
  cmake -E env CXXFLAGS="-g3 -O0" cmake -DgRPC_BUILD_TESTS=ON -DgRPC_SSL_PROVIDER=wolfssl ../..